fix big lint + populate user

+++ lib/net/meet_api.dart
@@ -12,24 +12,20 @@ class OppidumsMeetApi {
     return res.data;
   }
 
-    static Future<List<dynamic>> getOwnerMeet(String owner, String accessToken) async {
-    var res = await _client.get(
-      'meets',
-      queryParameters: {"owner": owner},
-      options: Options(headers: {'x-access-token': accessToken}));
-      
+  static Future<List<dynamic>> getOwnerMeet(String owner, String accessToken) async {
+    var res = await _client.get('meets',
+        queryParameters: {"owner": owner}, options: Options(headers: {'x-access-token': accessToken}));
+
     return res.data;
   }
 
   static Future<Map<String, dynamic>> getMeetById(String meetsId, String accessToken) async {
-    var res = await _client.get('meets/$meetsId',
-      options: Options(headers: {'x-access-token': accessToken}));
+    var res = await _client.get('meets/$meetsId', options: Options(headers: {'x-access-token': accessToken}));
     return res.data;
   }
-  
+
   static Future<Map<String, dynamic>> deleteMeetById(String meetsId, String accessToken) async {
-    var res = await _client.delete('meets/$meetsId',
-      options: Options(headers: {'x-access-token': accessToken}));
+    var res = await _client.delete('meets/$meetsId', options: Options(headers: {'x-access-token': accessToken}));
     return res.data;
   }
 
@@ -42,27 +38,23 @@ class OppidumsMeetApi {
     DateTime startDate,
     String accessToken,
   ) async {
-
-    var res = await _client.post('meets', data: {
-      'owner': userId,
-      'place': placeId,
-      'city': cityId,
-      'title': title,
-      'description': description,
-      'startDate': startDate.toString()
-    },
-      options: Options(headers: {'x-access-token': accessToken}));
+    var res = await _client.post('meets',
+        data: {
+          'owner': userId,
+          'place': placeId,
+          'city': cityId,
+          'title': title,
+          'description': description,
+          'startDate': startDate.toString()
+        },
+        options: Options(headers: {'x-access-token': accessToken}));
     return res.data;
   }
 
-  static Future<Map<String, dynamic>>  joinMeet(String meetId, List<dynamic> participens, String accessToken) async {
-    var res = await _client.put(
-      'meets/${meetId}',
-      data: {"participens": participens},
-      options: Options(headers: {'x-access-token': accessToken}));
+  static Future<Map<String, dynamic>> joinMeet(String meetId, List<dynamic> participens, String accessToken) async {
+    var res = await _client.put('meets/${meetId}',
+        data: {"participens": participens}, options: Options(headers: {'x-access-token': accessToken}));
 
     return res.data;
   }
-
-  
 }
